Use map[string]string for snapshot createopt
Snapshot create opts are using map[string]interface{} for the
metadata opt.  This is doable, but we should be consistent with
other objects/resources and use map[string]string.

I missed this on the initial commit, so this patch just modifies
the CreateOption struct to make it match the types we use in Volumes.