Cinder Storage Pools API (#205)
* Cinder Storage Pools API

Adds in the ability to probe storage backends to get the current and maximum
pool utilization.  Useful for capacity planning based on committed-to metrics
and those as a result of thin-provisioning.

* jtopjian's comments from 25/11/16

* Update field source documentation

Additionally makes the distinction between required and optional fields based
on said documentation, adding in the optional fields

* Fix up CI header format failures

* Move Storage Pools to Block Storage Extension

* Jrperritt's Review Comments

Cleans up wine-induced typos.  Updated the unmarshall code, seems the integer case is
unnecessarily as go when unmarshalling a number defaults to float64, so has been
deleted.  As we default to 0.0 for total and free capacities I've folded the unknown
case into the default.  Testing code has been updated to include an 'integer' in the
JSON to ensure the assumptions hold.

* Add Pagination

* Tweak Unmarshalling

Rather than unmarshal the entire storage pool struct we only need cater for
the capabilities structure, the JSON unmarshalling code auto-magically takes
care of invoking the correct routines for us.

* Update Name and Location

Renamed to scheduler stats, relocated to v2 extension as it's not part of v1.

* Move Back To Generic Extension

* Alter JSON Unmarshalling

Makes all defaults silently return 0.0 if something is amiss.  Also thought it best to
share the duplicated code in the form of a separate function.

* Remove Capacity Parser Global

Unbeknownst to me symbol visibility across compilation units is controlled
by capitalization, guess I'm too old and used to C's static :)  This removes
the export.  Jrperritt preferred the functionality in the actual JSON code, I
wanted to reduce code duplication so what I gather is the most idiomatic go
is function literals/lambdas.