Common resource pkg for Cinder V2 and V3
The baseline for Cinder's micro version enabled API (V3.0) is
a direct implementation of the current Cinder V2 API.  Rather
than try and create a specific pkg for each Cinder API version,
this PR proposes that instead we create a version agnostic pkg
for Cinder resources.  Things like specific micro versioned
requests and results can then be added/built on top of the common
baseline that is provided here if and when needed.